images/fonts

+++ ZincWallet/ZNPaymentRequest.h
@@ -13,7 +13,7 @@
 @property (nonatomic, strong) NSString *paymentAddress;
 @property (nonatomic, strong) NSString *label;
 @property (nonatomic, strong) NSString *message;
-@property (nonatomic, assign) double amount;
+@property (nonatomic, assign) uint64_t amount;
 @property (nonatomic, strong) NSData *data;
 @property (nonatomic, readonly) NSData *signedTransaction;
 @property (nonatomic, readonly, getter=isValid) BOOL valid;
